+++ src/main/java/com/oceandate/backend/domain/matching/dto/RotationResponse.java
@@ -2,8 +2,6 @@ package com.oceandate.backend.domain.matching.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,7 +10,7 @@ import java.math.BigDecimal;
 public class RotationResponse {
     private Long applicationId;
     private String orderId;
-    private BigDecimal amount;
+    private Integer amount;
     private String orderName;
     private String customerEmail;
 }
