+++ src/main/java/com/oceandate/backend/domain/matching/entity/RotationEvent.java
@@ -7,7 +7,6 @@ import com.oceandate.backend.global.exception.constant.ErrorCode;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -49,7 +48,7 @@ public class RotationEvent {
     private String location;
 
     @Column(nullable = false)
-    private BigDecimal amount;  // 참가비
+    private Integer amount;  // 참가비
 
     @Column(columnDefinition = "TEXT")
     private String description;
